+++ src/containers/FooterContainer/FooterContainer.vue
@@ -239,6 +239,10 @@ export default {
             {
               text: 'footer.press-kit.header',
               to: '/press-kit'
+            },
+            {
+              text: 'footer.generate-eth2',
+              to: '/generate-eth2-keystore'
             }
           ]
         },