+++ src/dapps/index.js
@@ -11,7 +11,9 @@ import ambrpayDisabled from '@/assets/images/icons/dapps/ambrpay-disabled.png';
 import aave from '@/assets/images/icons/dapps/aave.svg';
 import lendMigrator from '@/assets/images/icons/dapps/lend-migrator.svg';
 import unstoppable from '@/assets/images/icons/dapps/unstoppable.png';
+import staked from '@/assets/images/icons/dapps/staked-logo.svg';
 import golem from '@/assets/images/icons/dapps/Golem_Submark_Positive_RGB.svg';
+
 import { ETH, GOERLI, ROP, RIN, KOV } from '@/networks/types';
 
 const dapps = {
@@ -92,6 +94,15 @@ const dapps = {
     supportedNetworks: [ETH.name],
     lastUpdated: '10/2/2020',
     releaseDate: '10/2/2020'
+  },
+  staked: {
+    route: '/interface/dapps/staked',
+    icon: staked,
+    title: 'dappsStaked.title',
+    desc: 'dappsStaked.desc',
+    supportedNetworks: [ETH.name, GOERLI.name],
+    lastUpdated: '12/8/2020',
+    releaseDate: '12/8/2020'
   },
   gntMigrator: {
     route: '/interface/dapps/gnt-migrator',