+++ src/wallets/hardware/bcvault/errorHandler.js
@@ -19,8 +19,6 @@ const WARNING = {
 };
 
 export default err => {
-  // eslint-disable-next-line
-  console.log(err, JSON.stringify(err));
   // web errors
   if (err.hasOwnProperty('jsError')) {
     Toast.responseHandler(