+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Feature
 
 * Eth2 Staking Feature [#2662](https://github.com/MyEtherWallet/MyEtherWallet/pull/2662)
+
 ### Bug
 
 * Fixed various bugs from asana [#2693](https://github.com/MyEtherWallet/MyEtherWallet/pull/2693)