+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### Release v5.7.15
+
+### Feature
+
+* Eth2 Staking Feature [#2662](https://github.com/MyEtherWallet/MyEtherWallet/pull/2662)
+
+### Bug
+
+* Fixed various bugs from asana [#2693](https://github.com/MyEtherWallet/MyEtherWallet/pull/2693)
+* Fix error on print [#2676](https://github.com/MyEtherWallet/MyEtherWallet/pull/2676)
+
 ### Release v5.7.14
 
 ### Feature